+++ project_frontend/src/App.js
@@ -1,5 +1,5 @@
 import './App.css';
-import { Home, Login, Write } from './pages';
+import { Home, Login, Welcome, Write } from './pages';
 import { Routes, Route } from 'react-router-dom';
 
 function App() {
@@ -8,7 +8,7 @@ function App() {
 			<Routes>
 				<Route path="/" element={<Home />}/>
 				<Route path="/login" element={<Login />}/>
-				<Route path="/api/member/join/*" element={<Home />}/>
+				<Route path="/api/member/join/*" element={<Welcome />}/>
 				<Route path="/boards" element={<Write />}/>
 			</Routes>
 		</div>
